refactor: include new utilities in form r part a
